bzr revid: ls@numerigraphe.com-20130717103317-cf22a6614c4wzmv9

+++ addons/delivery/delivery.py
@@ -20,8 +20,9 @@
 ##############################################################################
 
 import time
-from osv import fields,osv
-from tools.translate import _
+from openerp.osv import fields,osv
+from openerp.tools.translate import _
+import openerp.addons.decimal_precision as dp
 
 class delivery_carrier(osv.osv):
     _name = "delivery.carrier"
@@ -68,7 +69,7 @@ class delivery_carrier(osv.osv):
         'price' : fields.function(get_price, string='Price'),
         'active': fields.boolean('Active', help="If the active field is set to False, it will allow you to hide the delivery carrier without removing it."),
         'normal_price': fields.float('Normal Price', help="Keep empty if the pricing depends on the advanced pricing per destination"),
-        'free_if_more_than': fields.boolean('Free If More Than', help="If the order is more expensive than a certain amount, the customer can benefit from a free shipping"),
+        'free_if_more_than': fields.boolean('Free If Order Total Amount Is More Than', help="If the order is more expensive than a certain amount, the customer can benefit from a free shipping"),
         'amount': fields.float('Amount', help="Amount of the order to benefit from a free shipping, expressed in the company currency"),
         'use_detailed_pricelist': fields.boolean('Advanced Pricing per Destination', help="Check this box if you want to manage delivery prices that depends on the destination, the weight, the total of the order, etc."),
         'pricelist_ids': fields.one2many('delivery.grid', 'carrier_id', 'Advanced Pricing'),
@@ -80,7 +81,7 @@ class delivery_carrier(osv.osv):
     }
 
     def grid_get(self, cr, uid, ids, contact_id, context=None):
-        contact = self.pool.get('res.partner.address').browse(cr, uid, contact_id, context=context)
+        contact = self.pool.get('res.partner').browse(cr, uid, contact_id, context=context)
         for carrier in self.browse(cr, uid, ids, context=context):
             for grid in carrier.grids_id:
                 get_id = lambda x: x.id
@@ -98,12 +99,17 @@ class delivery_carrier(osv.osv):
         return False
 
     def create_grid_lines(self, cr, uid, ids, vals, context=None):
-        if context == None:
+        if context is None:
             context = {}
         grid_line_pool = self.pool.get('delivery.grid.line')
         grid_pool = self.pool.get('delivery.grid')
         for record in self.browse(cr, uid, ids, context=context):
-            grid_id = grid_pool.search(cr, uid, [('carrier_id', '=', record.id),('sequence','=',9999)], context=context)
+            # if using advanced pricing per destination: do not change
+            if record.use_detailed_pricelist:
+                continue
+
+            # not using advanced pricing per destination: override grid
+            grid_id = grid_pool.search(cr, uid, [('carrier_id', '=', record.id)], context=context)
 
             if grid_id and not (record.normal_price or record.free_if_more_than):
                 grid_pool.unlink(cr, uid, grid_id, context=context)
@@ -112,13 +118,12 @@ class delivery_carrier(osv.osv):
                 continue
 
             if not grid_id:
-                record_data = {
+                grid_data = {
                     'name': record.name,
                     'carrier_id': record.id,
-                    'sequence': 9999,
+                    'sequence': 10,
                 }
-                new_grid_id = grid_pool.create(cr, uid, record_data, context=context)
-                grid_id = [new_grid_id]
+                grid_id = [grid_pool.create(cr, uid, grid_data, context=context)]
 
             lines = grid_line_pool.search(cr, uid, [('grid_id','in',grid_id)], context=context)
             if lines:
@@ -126,7 +131,7 @@ class delivery_carrier(osv.osv):
 
             #create the grid lines
             if record.free_if_more_than:
-                data = {
+                line_data = {
                     'grid_id': grid_id and grid_id[0],
                     'name': _('Free if more than %.2f') % record.amount,
                     'type': 'price',
@@ -135,10 +140,9 @@ class delivery_carrier(osv.osv):
                     'standard_price': 0.0,
                     'list_price': 0.0,
                 }
-                grid_line_pool.create(cr, uid, data, context=context)
-
+                grid_line_pool.create(cr, uid, line_data, context=context)
             if record.normal_price:
-                default_data = {
+                line_data = {
                     'grid_id': grid_id and grid_id[0],
                     'name': _('Default price'),
                     'type': 'price',
@@ -147,13 +151,15 @@ class delivery_carrier(osv.osv):
                     'standard_price': record.normal_price,
                     'list_price': record.normal_price,
                 }
-                grid_line_pool.create(cr, uid, default_data, context=context)
+                grid_line_pool.create(cr, uid, line_data, context=context)
         return True
 
     def write(self, cr, uid, ids, vals, context=None):
-        res_id = super(delivery_carrier, self).write(cr, uid, ids, vals, context=context)
+        if isinstance(ids, (int,long)):
+            ids = [ids]
+        res = super(delivery_carrier, self).write(cr, uid, ids, vals, context=context)
         self.create_grid_lines(cr, uid, ids, vals, context=context)
-        return res_id
+        return res
 
     def create(self, cr, uid, vals, context=None):
         res_id = super(delivery_carrier, self).create(cr, uid, vals, context=context)
@@ -212,7 +218,7 @@ class delivery_grid(osv.osv):
                 ok = True
                 break
         if not ok:
-            raise osv.except_osv(_('No price available!'), _('No line matched this product or order in the choosed delivery grid.'))
+            raise osv.except_osv(_('No price available!'), _('No line matched this product or order in the chosen delivery grid.'))
 
         return price
 
@@ -223,7 +229,7 @@ class delivery_grid_line(osv.osv):
     _name = "delivery.grid.line"
     _description = "Delivery Grid Line"
     _columns = {
-        'name': fields.char('Name', size=32, required=True),
+        'name': fields.char('Name', size=64, required=True),
         'grid_id': fields.many2one('delivery.grid', 'Grid',required=True, ondelete='cascade'),
         'type': fields.selection([('weight','Weight'),('volume','Volume'),\
                                   ('wv','Weight * Volume'), ('price','Price')],\
@@ -232,8 +238,8 @@ class delivery_grid_line(osv.osv):
         'max_value': fields.float('Maximum Value', required=True),
         'price_type': fields.selection([('fixed','Fixed'),('variable','Variable')], 'Price Type', required=True),
         'variable_factor': fields.selection([('weight','Weight'),('volume','Volume'),('wv','Weight * Volume'), ('price','Price')], 'Variable Factor', required=True),
-        'list_price': fields.float('Sale Price', required=True),
-        'standard_price': fields.float('Cost Price', required=True),
+        'list_price': fields.float('Sale Price', digits_compute= dp.get_precision('Product Price'), required=True),
+        'standard_price': fields.float('Cost Price', digits_compute= dp.get_precision('Product Price'), required=True),
     }
     _defaults = {
         'type': lambda *args: 'weight',
@@ -244,24 +250,5 @@ class delivery_grid_line(osv.osv):
     _order = 'list_price'
 
 delivery_grid_line()
-
-class define_delivery_steps(osv.osv_memory):
-    _name = 'delivery.define.delivery.steps.wizard'
-
-    _columns = {
-        'picking_policy' : fields.selection([('direct', 'Deliver each product when available'), ('one', 'Deliver all products at once')], 'Picking Policy'),
-    }
-    _defaults = {
-        'picking_policy': lambda s,c,u,ctx: s.pool.get('sale.order').default_get(c,u,['picking_policy'],context=ctx)['picking_policy']
-    }
-
-    def apply_cb(self, cr, uid, ids, context=None):
-        ir_values_obj = self.pool.get('ir.values')
-        wizard = self.browse(cr, uid, ids, context=context)[0]
-        ir_values_obj.set(cr, uid, 'default', False, 'picking_policy', ['sale.order'], wizard.picking_policy)
-        return {'type' : 'ir.actions.act_window_close'}
-
-define_delivery_steps()
-
 
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
